+++ CHANGELOG.md
@@ -819,6 +819,7 @@
 - Added a new accepted channel of delivery code, "22000" ("Donor country-based NGO")
 - Force word wrapping in table cells showing invalid upload values
 - Allow a BEIS user to export external income for all delivery partner organisations
+- Countries and regions are derived from BEIS benefitting countries code lists
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-72...HEAD
 [release-72]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-71...release-72